Add files via upload
This is a dashboard created using PowerBI. The dash is a geographical catalogue of the meteorites held in the Natural History Museum Collection. The dashboard is dynamic and data visualised changes based on user input. A slicer was included to allow users to reduce data based on year. The user can also drill through to further details about each of the meteorites using the map and the drill through button at the top of the page. 

Data Source: Monica Grady (2022). Catalogue of Meteorites [Data set]. Natural History Museum. https://doi.org/10.5519/tqfuwle7