Add files via upload
This is a PowerBI dashboard I created looking at tourist visits to London between 2002 and 2020. I initially started with a dataset from the London Authority looking at tourist visits, spends, reasons for visiting and countries of origin. I built on this by getting additional visitor information and geographical data for some of the top London tourist attractions. 
To make this dynamic, I focused on using slicers to change the dashboard on user input and customised map tooltips. I also added buttons to navigate between the different sheets and used bar and line visualisations to show multiple linked information in graphs.